LayoutTests/imported/w3c/web-platform-tests/web-animations/animation-model/keyframe-effects/effect-value-context-expected.txt


FAIL Effect values reflect changes to font-size on element assert_equals: Effect value before updating font-size expected "150px" but got "0px"
FAIL Effect values reflect changes to font-size on parent element assert_equals: Effect value before updating font-size on parent element expected "150px" but got "0px"
FAIL Effect values reflect changes to font-size when computed style is not immediately flushed assert_equals: Effect value after updating font-size on parent element expected "300px" but got "0px"
FAIL Effect values reflect changes to font-size from reparenting assert_equals: Effect value after attaching to font-size:10px parent expected "150px" but got "0px"
FAIL Effect values reflect changes to target element assert_equals: Effect value before updating target element expected "150px" but got "0px"
